Refactor: ScrollView Transition 리팩토링
